refactor: move pure bench result types to homeboy-extension-contract
Stage-2 cut into the bench subsystem - the largest reverse-edge cluster
(bench = 58 core prod refs). Extracts the 14 pure result types that have no
observation/finding/sibling-behavior dependencies:

RigPackageEvidence, RigPackageFreshness, BenchRunExecution, BenchProvenance
(+ link), BenchWorkloadMetadata, BenchRunnerMetadata, BenchChildCommandFailure,
BenchMetrics, BenchMetricPolicy, BenchMetricDirection, BenchMetricPhase,
RegressionTest, BenchMemory - plus their trivial pure impls (BenchProvenance
is_empty, BenchMetrics get) and serde helpers (provenance_is_empty, is_false).

The entangled result types (BenchResults, BenchRunMetadata, BenchScenario,
BenchRunSnapshot) stay in core - they reference HomeboyFinding,
LifecycleResultMetadata, observation::timeline types, and other bench
submodules (artifact/diagnostic/gate/phase_events). core's result_types.rs
re-imports the moved types, keeping crate::extension::bench::* paths stable.

This clears two core-imported bench symbols (RigPackageEvidence,
BenchRunExecution) toward the eventual homeboy-extension crate. The remaining
bench entanglement is gated on an observation::timeline relocation (that file is
already core-dep-free, 685 LOC - a future untangle candidate).
